requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
pydantic-settings
xhtml2pdf
SQLAlchemy
mysql-connector-python
asyncmy
python-docx
html2docx
markdown2
json-repair
openai
pytest
httpx
thefuzz